requirements.txt
# Core
torch==2.6.0
torchvision==0.21.0
lightning==2.6.0
numpy==1.24.4

# Config
hydra-core==1.3.2
omegaconf==2.3.0
pyyaml==6.0.2

# Model stack
transformers==4.51.3
diffusers==0.34.0
accelerate==1.1.1
tokenizers==0.21.4
sentencepiece==0.2.0
ftfy==6.3.1
einops==0.8.0
peft==0.15.0

# IO
pandas==2.1.4
pillow==11.1.0
decord==0.6.0
imageio==2.34.0
imageio-ffmpeg==0.5.1
av==14.0.1

# Logging
wandb==0.19.1
wandb-osh==1.2.2


# Utilities
tqdm==4.66.2
colorama==0.4.6
msgpack==1.1.2
pyzmq==27.1.0
